requirements.txt
fastapi[standard]==0.118.0
grpcio==1.75.1
grpcio-tools==1.75.1
uvicorn==0.37.0